gui: Display source files using google-code-prettify
WebView seems to be the best approach for display source code going forward,
this is also an established, working, method for syntax highlighting.
Though, in the future I would like to totally avoid JavaScript here. There
are benefits to moving to WebView:
1. There is FAR more available tooling for the web than RichTextFX, reducing
   our work.
2. It is very easy to manipulate a website, unless the CodeView from
   RichTextFX - this meant it was likely (as happened with Nocturne) we would
   have had to move away from it anyhow.

Another concern is that I'm using a CDN to get the library, leaving a subpar
experience for those without an internet connection - this will need to be
resolved! The issue is that google-code-prettify isn't versioned, and so a
regular webjar won't suffice :( I see a few solutions to this, which I will
be considering in the coming days/weeks:
1. Include their distribution as part of our build process, downloading a
   their distribution at a specified commit.
2. Have a seperate project to handle Java distribution of
   google-code-prettify, that we would consume.
3. Move to another code prettifier, that is versioned.